Globalization
Made almost all variables global. This makes it so you can call them
without instantiating them anywhere, and it also got rid of the huge
parameter list for the methods. Also put the voting into its own method
called voting. Call it in the rounds for easy insertion and for sake of
time.